Split packing interface and internals files
