+++ source/llassetgen/CMakeLists.txt
@@ -33,6 +33,9 @@ set(include_path "${CMAKE_CURRENT_SOURCE_DIR}/include/${target}")
 set(source_path  "${CMAKE_CURRENT_SOURCE_DIR}/source")
 
 set(headers
+    ${include_path}/packing/internal/ShelfPacker.h
+    ${include_path}/packing/Algorithms.h
+    ${include_path}/packing/Types.h
     ${include_path}/llassetgen.h
     ${include_path}/DistanceTransform.h
     ${include_path}/Geometry.h
@@ -42,7 +45,7 @@ set(headers
 set(sources
     ${source_path}/llassetgen.cpp
     ${source_path}/DistanceTransform.cpp
-    ${source_path}/Packing.cpp
+    ${source_path}/packing/internal/ShelfPacker.cpp
 )
 
 # Group source files
